chore: reorganize the header component and its sub components
