--- client/src/components/common/Header/components/index.ts
@@ -1,4 +1,0 @@
-export { default as DateDropdown } from "./DateDropdown";
-export { default as NotificationBell } from "./NotificationBell";
-export { default as SearchInput } from "./SearchInput";
-export { default as MobileMenu } from "./MobileMenu";
