--- client/src/components/common/Header/index.ts
@@ -1,1 +1,0 @@
-export { default as DashboardHeader } from "./Header";
